LoginScreen with POST functionality
LoginScreen with POST functionality. Sends a POST request to the specified API endpoint. If the email and password entered matches with the information on the endpoint, u will be logged in and a JSON Web Token will be created. This token lasts for 1000seconds before it expires. The user will then be logged out.